Use consistent agency contact information
There are four locations where agency contact information appears and they
should all be identical.
