+++ js/components/agency_component_preview.jsx
@@ -2,22 +2,16 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import AgencyComponentProcessingTime from './agency_component_processing_time';
-import FoiaPersonnel from './foia_personnel';
-import FoiaSubmissionAddress from './foia_submission_address';
+import ContactInformation from './contact_information';
 import NonInteroperableInfo from './non_interoperable_info';
 import { AgencyComponent } from '../models';
 import domify from '../util/request_form/domify';
-import foiaPersonnel from '../util/foia_personnel';
 
 
 function AgencyComponentPreview({ onAgencySelect, agencyComponent, isCentralized }) {
   const description = AgencyComponent.agencyMission(agencyComponent);
   const requestUrl = `/request/agency-component/${agencyComponent.id}/`;
   const onSelect = () => onAgencySelect(agencyComponent.agency);
-
-  // Grab contacts
-  const serviceCenterFoiaPersonnel = foiaPersonnel.personnel(agencyComponent, 'service_centers')[0];
-  const publicLiaisonFoiaPersonnel = foiaPersonnel.personnel(agencyComponent, 'public_liaisons')[0];
 
   return (
     <div className="agency-preview usa-grid-full">
@@ -50,28 +44,8 @@ function AgencyComponentPreview({ onAgencySelect, agencyComponent, isCentralized
           </div>
         }
 
-        <h4>
-          <span data-term="foia requester service center" >FOIA Requester Service Center</span>
-        </h4>
-        <div className="agency-preview_contact-section">
-          <FoiaPersonnel foiaPersonnel={agencyComponent.service_centers[0]} />
-        </div>
-        <h4><span data-term="foia public liaison" >FOIA Public Liaison</span></h4>
-        <div className="agency-preview_contact-section">
-          <FoiaPersonnel foiaPersonnel={serviceCenterFoiaPersonnel} />
-        </div>
-        <div className="agency-preview_contact-section">
-          <FoiaPersonnel foiaPersonnel={publicLiaisonFoiaPersonnel} />
-        </div>
-        <div className="agency-preview_contact-section">
-          <FoiaSubmissionAddress submissionAddress={agencyComponent.submission_address} />
-          { agencyComponent.email &&
-            <p className="agency-info_email">
-              <a href={`mailto:${agencyComponent.email}`}>{ agencyComponent.email }</a>
-            </p>
-          }
-        </div>
-
+        <h4>Contact</h4>
+        <ContactInformation agencyComponent={agencyComponent} />
       </div>
       <div className="usa-width-one-half start-request-container">
         { agencyComponent.request_data_year &&
