+++ www.foia.gov/_sass/_modules.scss
@@ -2,6 +2,7 @@
 @import 'modules/agency_preview';
 @import 'modules/buttons';
 @import 'modules/confirmation';
+@import 'modules/contact_information';
 @import 'modules/developer';
 @import 'modules/feedback_bar';
 @import 'modules/foia_request_form';
