+++ www.foia.gov/_sass/modules/_agency_preview.scss
@@ -33,11 +33,6 @@
   list-style: none;
 }
 
-.agency-preview_contact-section {
-  color: $color-primary-darkest;
-  margin-bottom: $space-4x;
-}
-
 .agency-preview_list-hed {
   @extend %h2;
 }
