+++ www.foia.gov/_sass/modules/_contact_information.scss
@@ -1,0 +1,8 @@
+.contact-information_section {
+  margin-bottom: $space-4x;
+
+  p {
+    margin-bottom: $space-1x;
+    margin-top: $space-1x;
+  }
+}
